requirements.txt
streamlit==1.41.1
google-generativeai>=0.8.0
beautifulsoup4==4.12.3
requests==2.32.3
yt-dlp==2025.01.26
python-dotenv==1.0.1
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
pydantic>=2.0.0
langchain-chroma>=0.1.0
langchain-community>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
prometheus_client>=0.20.0
duckduckgo-search>=6.0.0
fpdf2>=2.7.0
Pillow>=10.0.0
